wip database tables & request/response models
