+++ miscellaneous/database/db.py
@@ -4,4 +4,5 @@ db.py
 from sqlmodel import create_engine
 from settings import settings
 
+
 engine = create_engine(settings.database, echo=True)
